+++ ChangeLog.md
@@ -4,6 +4,10 @@
 - **PlatformSwitch Enhancements**:
   - Upgraded to fully integrate with `sharedPolling`.
   - Added support for the `getNestedValue` function to process nested JSON data.
+- **PlatformSensor Enhancements**:
+  - Upgraded to fully integrate with `sharedPolling`.
 - **PlatformCarbonDioxide Class**:
   - Added a new class `platformCarbonDioxide` to support devices monitoring carbon dioxide levels.
   - Integrated a new service for managing and reporting CO₂ data.
+  - Support Discord - Detect CO2, Low Battery
+  - Added support for the `getNestedValue` function to process nested JSON data.